// src/Types.ts
export type RawErrorType =
  | 'invalid_request_error'
  | 'api_error'
  | 'rate_limit_error'
  | 'authentication_error';
// ...
export type LotrRawError = {
  message?: string;
  type?: RawErrorType;
  headers?: {[header: string]: string};
  statusCode?: number;
  requestId?: string;
  code?: string;
  doc_url?: string;
  param?: string;
  detail?: string | Error;
  source?: any;
  exception?: any;
};